tests: more coverage
